requirements.txt
starlette
pyheif
httpx
Pillow
boto3
python-dotenv
